Moved test_support out of utilities
